Add support for deployments initially scaled to zero
backport of https://github.com/deislabs/osiris/pull/40

When using a deployment initially scaled to zero replicas, Osiris won't automatically scale it up when needed.
This is because the associated service is not updated with the right endpoints until there is a least 1 pod event.

To fix it I just added an initial call to the `syncEndpoints` func when running the endpoints manager,
so that the service will be initially configured with Osiris activator endpoint.