fix: app activation timeout was too short
the app activation was linked to the request that initiated the activation. so if the request was cancelled after 1 second, an app activation with a dependency would fail to activate the real app.
we're using an hardcoded timeout of 5 minutes for the moment. if needed we'll make it configurable later